requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
Django==5.2.1
djangorestframework==3.16.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
httplib2==0.22.0
idna==3.10
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic_core==2.33.2
pyparsing==3.2.3
python-decouple==3.8
requests==2.32.3
rsa==4.9.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
